MAINT(refac): streamline use to avoid code duplication
Add loader and plotter registries, as `dict` with keys corresponding to
the matcher `enum`. These registries are filled on class declaration
with some `attrs` magic, so that their values hold the corresponding
classes, and a call to `dict[key]` can be used for class instantiation,
avoiding some repetitive switch-like constructions.